+++ src/iterator.rs
@@ -3,24 +3,24 @@ use crate::intervals::Interval;
 use crate::nothing_between::NothingBetween;
 use crate::step::Step;
 
-pub struct IntervalIterator<T>
-where
-    T: Step + Clone + PartialOrd + NothingBetween,
-{
+pub struct IntervalIterator<T> {
     pub(crate) intv: Interval<T>,
 }
 
-impl<T> IntervalIterator<T>
-where
-    T: Step + Clone + PartialOrd + NothingBetween,
-{
+impl<T> IntervalIterator<T> {
     /// Return an interval matching what the iterators will return
-    pub fn as_interval(&self) -> Interval<T> {
+    pub fn as_interval(&self) -> Interval<T>
+    where
+        T: Clone,
+    {
         self.intv.clone()
     }
 
     /// Internal implementation for nth() and next()
-    fn internal_nth(&mut self, n: usize) -> Option<T> {
+    fn internal_nth(&mut self, n: usize) -> Option<T>
+    where
+        T: Step + Clone + PartialOrd + NothingBetween,
+    {
         if self.intv.is_empty() {
             return None;
         }
@@ -59,7 +59,10 @@ where
     }
 
     /// Internal implementation for nth_back() and next_back()
-    fn internal_nth_back(&mut self, n: usize) -> Option<T> {
+    fn internal_nth_back(&mut self, n: usize) -> Option<T>
+    where
+        T: Step + Clone + PartialOrd + NothingBetween,
+    {
         if self.intv.is_empty() {
             return None;
         }
@@ -112,6 +115,51 @@ where
     fn nth(&mut self, n: usize) -> Option<Self::Item> {
         self.internal_nth(n)
     }
+
+    /// Used to compute the result of `ExactSizeIterator::len()`, and
+    /// optimize calls to collect() by pre-allocating when possible.
+    fn size_hint(&self) -> (usize, Option<usize>) {
+        if self.intv.is_empty() {
+            (0, Some(0))
+        } else {
+            let len = match (&self.intv.lower, &self.intv.upper) {
+                (Bound::RightUnbounded, _) | (_, Bound::LeftUnbounded) => {
+                    Some(0)
+                }
+                (Bound::LeftUnbounded, Bound::RightUnbounded) => {
+                    T::min_value().elements_between(&T::max_value())
+                }
+                (Bound::LeftUnbounded, Bound::LeftOf(up)) => {
+                    T::min_value().elements_between(up)
+                }
+                (Bound::LeftUnbounded, Bound::RightOf(up)) => {
+                    T::min_value().elements_between(up).map(|c| c + 1)
+                }
+                (Bound::LeftOf(lo), Bound::RightUnbounded) => {
+                    lo.elements_between(&T::max_value())
+                }
+                (Bound::LeftOf(lo), Bound::LeftOf(up)) => {
+                    lo.elements_between(up)
+                }
+                (Bound::LeftOf(lo), Bound::RightOf(up)) => {
+                    lo.elements_between(up).map(|c| c + 1)
+                }
+                (Bound::RightOf(lo), Bound::RightUnbounded) => {
+                    lo.elements_between(&T::max_value()).map(|c| c - 1)
+                }
+                (Bound::RightOf(lo), Bound::LeftOf(up)) => {
+                    lo.elements_between(up).map(|c| c - 1)
+                }
+                (Bound::RightOf(lo), Bound::RightOf(up)) => {
+                    lo.elements_between(up)
+                }
+            };
+            match len {
+                None => (usize::MAX, None),
+                Some(l) => (l, Some(l)),
+            }
+        }
+    }
 }
 
 impl<T> DoubleEndedIterator for IntervalIterator<T>
@@ -126,4 +174,11 @@ where
     fn nth_back(&mut self, n: usize) -> Option<T> {
         self.internal_nth_back(n)
     }
+}
+
+/// len() will panic! if the number of values in the range is
+/// greater than usize::MAX.
+impl<T> ExactSizeIterator for IntervalIterator<T> where
+    T: Step + Clone + PartialOrd + NothingBetween
+{
 }